requirements.txt
easyocr>=1.7.0
flask>=3.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
werkzeug>=3.0.0
torch>=2.0.0
torchvision>=0.15.0
